File directory overhaul
Reworked app to use packages instead of being one giant module
