+++ runserver.py
@@ -1,0 +1,3 @@
+from summonercolor import app
+
+app.run()
